pyproject.toml
[tool.poetry]
name = "haha"
version = "0.1.0"
description = ""
authors = ["wjdtjddns98 <wjdtjddns98@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.2.3"
djangorestframework = "^3.16.0"
pyjwt = "^2.9.0"
djangorestframework-simplejwt = "^5.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
